requirements.txt
certifi==2024.2.2
gevent==24.2.1
greenlet==3.0.3
MouseInfo==0.1.3
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==10.1
pyobjc-framework-Cocoa==10.1
pyobjc-framework-Quartz==10.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dotenv==1.0.1
pytweening==1.0.7
rubicon-objc==0.4.7
setuptools==69.1.1
websocket-client==1.7.0
websockets==12.0
zope.event==5.0
zope.interface==6.2
